all translations on one page and minimalized options

+++ src/Controller.php
@@ -1,23 +1,29 @@
 <?php namespace Barryvdh\TranslationManager;
 
 use Illuminate\Http\Request;
-use Illuminate\Routing\Controller as BaseController;
 use Barryvdh\TranslationManager\Models\Translation;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 
-class Controller extends BaseController
+class Controller extends \App\Http\Controllers\Controller
 {
     /** @var \Barryvdh\TranslationManager\Manager  */
     protected $manager;
 
     public function __construct(Manager $manager)
     {
+        $this->middleware('auth');
+        parent::__construct();
+
         $this->manager = $manager;
     }
 
     public function getIndex($group = null)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $locales = $this->manager->getLocales();
         $groups = Translation::groupBy('group');
         $excludedGroups = $this->manager->getConfig('exclude_groups');
@@ -32,6 +38,18 @@ class Controller extends BaseController
         $groups = [''=>'Choose a group'] + $groups;
         $numChanged = Translation::where('group', $group)->where('status', Translation::STATUS_CHANGED)->count();
 
+        $allTranslationsByGroup = [];
+        foreach($groups as $singleGroup) {
+            $allTranslations = Translation::where('group', $singleGroup)->orderBy('key', 'asc')->get();
+            $numTranslations = count($allTranslations);
+            $allTranslationsByGroup[$singleGroup] = [
+                'editUrl' => action('\Barryvdh\TranslationManager\Controller@postEdit', [$singleGroup]),
+            ];
+            foreach($allTranslations as $translation){
+                $allTranslationsByGroup[$singleGroup]['translations'][$translation->key][$translation->locale] = $translation;
+            }
+
+        }
 
         $allTranslations = Translation::where('group', $group)->orderBy('key', 'asc')->get();
         $numTranslations = count($allTranslations);
@@ -40,10 +58,11 @@ class Controller extends BaseController
             $translations[$translation->key][$translation->locale] = $translation;
         }
 
-         return view('translation-manager::index')
+        return view('translation-manager::index')
             ->with('translations', $translations)
             ->with('locales', $locales)
             ->with('groups', $groups)
+            ->with('allTranslationsByGroup', $allTranslationsByGroup)
             ->with('group', $group)
             ->with('numTranslations', $numTranslations)
             ->with('numChanged', $numChanged)
@@ -53,11 +72,19 @@ class Controller extends BaseController
 
     public function getView($group = null)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         return $this->getIndex($group);
     }
 
     protected function loadLocales()
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         //Set the default locale as the first one.
         $locales = Translation::groupBy('locale')
             ->select('locale')
@@ -73,6 +100,10 @@ class Controller extends BaseController
 
     public function postAdd($group = null)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $keys = explode("\n", request()->get('keys'));
 
         foreach($keys as $key){
@@ -86,6 +117,10 @@ class Controller extends BaseController
 
     public function postEdit($group = null)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         if(!in_array($group, $this->manager->getConfig('exclude_groups'))) {
             $name = request()->get('name');
             $value = request()->get('value');
@@ -105,6 +140,10 @@ class Controller extends BaseController
 
     public function postDelete($group, $key)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         if(!in_array($group, $this->manager->getConfig('exclude_groups')) && $this->manager->getConfig('delete_enabled')) {
             Translation::where('group', $group)->where('key', $key)->delete();
             return ['status' => 'ok'];
@@ -113,6 +152,10 @@ class Controller extends BaseController
 
     public function postImport(Request $request)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $replace = $request->get('replace', false);
         $counter = $this->manager->importTranslations($replace);
 
@@ -121,6 +164,10 @@ class Controller extends BaseController
 
     public function postFind()
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $numFound = $this->manager->findTranslations();
 
         return ['status' => 'ok', 'counter' => (int) $numFound];
@@ -128,7 +175,11 @@ class Controller extends BaseController
 
     public function postPublish($group = null)
     {
-         $json = false;
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
+        $json = false;
 
         if($group === '_json'){
             $json = true;
@@ -141,6 +192,10 @@ class Controller extends BaseController
 
     public function postAddGroup(Request $request)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $group = str_replace(".", '', $request->input('new-group'));
         if ($group)
         {
@@ -154,6 +209,10 @@ class Controller extends BaseController
 
     public function postAddLocale(Request $request)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $locales = $this->manager->getLocales();
         $newLocale = str_replace([], '-', trim($request->input('new-locale')));
         if (!$newLocale || in_array($newLocale, $locales)) {
@@ -165,6 +224,10 @@ class Controller extends BaseController
 
     public function postRemoveLocale(Request $request)
     {
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         foreach ($request->input('remove-locale', []) as $locale => $val) {
             $this->manager->removeLocale($locale);
         }
@@ -172,6 +235,10 @@ class Controller extends BaseController
     }
 
     public function postTranslateMissing(Request $request){
+        if(!$this->current_user->hasRole('superadmin')) {
+            abort(401);
+        }
+
         $locales = $this->manager->getLocales();
         $newLocale = str_replace([], '-', trim($request->input('new-locale')));
         if($request->has('with-translations') && $request->has('base-locale') && in_array($request->input('base-locale'),$locales) && $request->has('file') && in_array($newLocale, $locales)){
